Now new and improved with support for booting stock kernel and SafeStrap
We now by default add a boot option for stock kernel and SafeStrap.
We do this by adding /sbin/preinit.sh script that starts kexecboot
as pid 1 so we can pivot_root to the stock Android distro.

We also enable uart support and set the default root password with:

Note that also utags-mmcblk1p13.bin needs to be reflashed as it now
starts /sbin/preinit.sh.

Signed-off-by: Tony Lindgren <tony@atomide.com>