Update with initial support for xt910, mz609, and mz617
Please note the new naming for the utags-*.bin files with the utags
partition name and droid4-kexecboot partition name.

If somebody has allow-mbmloader-flashing-mbm.bin for mz617, please
upload it somewhere and let me know. It seems to be only available
online right now for mz609.

Note that without allow-mbmloader-flashing-mbm.bin mz617 must be
flashed with dd for utags so there's a chance of bricking it without
a way to recover unless allow-mbmloader-flashing-mbm.bin is found.

Signed-off-by: Tony Lindgren <tony@atomide.com>